+++ libs/components/lists/src/lib/modules/repeater/fixtures/a11y-repeater-item.ts
@@ -1,0 +1,11 @@
+export interface A11yRepeaterItem {
+  selectable?: boolean;
+  selected?: boolean;
+  disabled?: boolean;
+  expanded?: boolean;
+  reorderable?: boolean;
+  title?: string;
+  message?: string;
+  context?: boolean;
+  tag?: string;
+}
